Fix format string confusion that was somehow related to r15510.
git-svn-id: svn://svn.valgrind.org/valgrind/trunk@15556

+++ coregrind/m_syswrap/syswrap-arm64-linux.c
@@ -442,7 +442,7 @@ PRE(sys_mmap)
    SysRes r;
 
    PRINT("sys_mmap ( %#lx, %lu, %lu, %#lx, %lu, %lu )",
-         ARG1, (ULong)ARG2, ARG3, ARG4, ARG5, ARG6 );
+         ARG1, ARG2, ARG3, ARG4, ARG5, ARG6 );
    PRE_REG_READ6(long, "mmap",
                  unsigned long, start, unsigned long, length,
                  unsigned long, prot,  unsigned long, flags,
